endianness + 64 bit issue's design proposed.

+++ src/mem.c
@@ -6,13 +6,13 @@
 #endif
 
 void *
-li_malloc(unsigned long int size)
+li_malloc(size_t size)
 {
     void *p;
 
     if (size + stats.mem_used > (settings.mem_avail)) {
         //syslog(LOG_ERR, "No memory available![%u MB]", settings.mem_avail);
-        dprintf("No memory available! %lu, %u, %lu", settings.mem_avail, stats.mem_used, size);
+        dprintf("No memory available! %lu, %u, %u", settings.mem_avail, stats.mem_used, size);
         return NULL;
     }
