+++ src/mem.h
@@ -4,7 +4,7 @@
 #ifndef MEM_H
 #define MEM_H
 
-void *li_malloc(unsigned long int size);
+void *li_malloc(size_t size);
 void li_free(void *ptr);
 
 #endif
